Add SingleProjectByNameService
