+++ src/main/java/sv/edu/ues/webhook/services/SingleProjectByNameService.java
@@ -1,48 +1,49 @@
 package sv.edu.ues.webhook.services;
 
 import com.fasterxml.jackson.databind.JsonNode;
-import com.google.api.services.dialogflow.v2beta1.model.GoogleCloudDialogflowV2IntentMessage;
 import com.google.api.services.dialogflow.v2beta1.model.GoogleCloudDialogflowV2WebhookResponse;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Value;
-import org.springframework.stereotype.Service;
 import org.springframework.web.client.HttpClientErrorException;
 import org.springframework.web.client.RestTemplate;
 import org.springframework.web.util.UriComponentsBuilder;
-import sv.edu.ues.webhook.annotations.IntentHandler;
 import sv.edu.ues.webhook.utils.PayloadBuilder;
 
-import java.util.List;
 import java.util.Map;
 
-@Service
-public class CertificadoService implements ExternalResourcesHandler {
+public class SingleProjectByNameService implements ExternalResourcesHandler{
+
+    private final Logger logger = LoggerFactory.getLogger(getClass());
+
 
     @Value("${external_resource.url_base}")
     private String baseUrl;
-    private String carnet, projectName;
     private final RestTemplate client;
-    private final Logger logger = LoggerFactory.getLogger(getClass());
+    private String carnet;
+    private String projectName;
 
-    public CertificadoService(RestTemplate client) {
-        this.client = client;
+    public SingleProjectByNameService(RestTemplate template) {
+        this.client = template;
     }
 
     @Override
     public String getExternalResourceUrl() {
         return UriComponentsBuilder.fromUriString(baseUrl)
-                .path("certificados/")
+                .path("estudiantes/")
                 .path(carnet)
+                .path("/proyectos")
                 .queryParam("projectName", projectName)
                 .toUriString();
     }
 
     @Override
     public void externalCall(GoogleCloudDialogflowV2WebhookResponse response) {
+        var uri = this.getExternalResourceUrl();
+
         JsonNode clientResponse;
         try {
-            clientResponse = this.client.getForObject(this.getExternalResourceUrl(), JsonNode.class);
+            clientResponse = this.client.getForObject(uri, JsonNode.class);
         }catch (HttpClientErrorException e){
             logger.error(e.getMessage());
             response.setFulfillmentText("Algo anda mal, por favor intente en unos minutos");
@@ -52,20 +53,17 @@ public class CertificadoService implements ExternalResourcesHandler {
 
         var message = clientResponse.get("message");
         if(message == null) {
-            var payload = PayloadBuilder.buildForAttachment(clientResponse.get("uri"));
-
-            var messages = new GoogleCloudDialogflowV2IntentMessage();
-            messages.setPayload(payload).setPlatform(PLATFORM);
-            response.setFulfillmentMessages(List.of(messages));
+            var builder = new StringBuilder();
+            PayloadBuilder.buildForProjectInfo(clientResponse, builder);
+            response.setFulfillmentText(builder.toString());
         }
         else
             response.setFulfillmentText(message.asText());
     }
 
     @Override
-    @IntentHandler(intent = "Certificado")
     public void handle(GoogleCloudDialogflowV2WebhookResponse response, Map<String, Object> params) {
-        logger.info("Resolving request for Certificado, current params: {}", params);
+        logger.info("Resolving request for SingleProjectByNameService, current params: {}", params);
         carnet = (String) params.get("carnet");
         projectName = (String) params.get("proyecto");
         if(!carnet.isBlank() && !projectName.isBlank())
